+++ src/FileManager/adapters/ExpressHttpAdapter.ts
@@ -1,0 +1,50 @@
+import { downloadSingleFile } from '../Interactor';
+import { Router, Response, Request } from 'express';
+import { LEARNING_OBJECT_ROUTES } from '../../shared/routes';
+import { mapErrorToResponseData } from '../../shared/errors';
+import { fileNotFound } from '../assets/filenotfound';
+
+export function buildRouter(): Router {
+  const router = Router();
+  router
+    .get('/users/:username/learning-objects/:loId/files/:fileId/download', download);
+  return router;
+}
+
+async function download(req: Request, res: Response) {
+  try {
+    const open = req.query.open;
+    const author: string = req.params.username;
+    const learningObjectId: string = req.params.loId;
+    const fileId: string = req.params.fileId;
+    const { filename, mimeType, stream } = await downloadSingleFile({
+      author,
+      fileId,
+      learningObjectId,
+    });
+    if (!open) {
+      res.attachment(filename);
+    }
+    // Set mime type only if it is known
+    if (mimeType) res.contentType(mimeType);
+    stream.pipe(res);
+  } catch (e) {
+    if (e.message === 'File not found') {
+      fileNotFoundResponse(e.object, req, res);
+    } else {
+      const { code, message } = mapErrorToResponseData(e);
+      res.status(code).json({message});
+    }
+  }
+}
+
+function fileNotFoundResponse(object: any, req: Request, res: Response) {
+  const redirectUrl = LEARNING_OBJECT_ROUTES.CLARK_DETAILS({
+    objectName: object.name,
+    username: req.params.username,
+  });
+  res
+    .status(404)
+    .type('text/html')
+    .send(fileNotFound(redirectUrl));
+}
