+++ src/FileManager/adapters/index.ts
@@ -1,0 +1,3 @@
+import * as ExpressHttpAdapter from './ExpressHttpAdapter';
+
+export { ExpressHttpAdapter };
