+++ src/FileManager/config/aws-sdk.config.ts
