+++ src/LearningObjects/PDFKitDriver.ts
@@ -1,7 +1,7 @@
 import * as PDFKit from 'pdfkit';
 import { LEARNING_OBJECT_ROUTES } from '../shared/routes';
 import { FileManager } from '../shared/interfaces/interfaces';
-import { FileUpload } from '../shared/interfaces/FileManager';
+import { FileUpload } from '../FileManager/interfaces/FileManager';
 import * as striptags from 'striptags';
 import { LearningObject, LearningOutcome } from '../shared/entity';
