+++ src/drivers/drivers.ts
@@ -3,7 +3,7 @@ import { ExpressRouteDriver } from './express/ExpressRouteDriver';
 import { ExpressAuthRouteDriver } from './express/ExpressAuthRouteDriver';
 import { ExpressAdminRouteDriver } from './express/ExpressAdminRouteDriver';
 import { MongoDriver } from './MongoDriver';
-import { S3Driver } from '../FileManager/S3Driver';
+import { S3Driver } from '../FileManager/drivers/S3Driver';
 export {
   ExpressDriver,
   ExpressRouteDriver,
